[EGD-7495] Fixed log use on Linux to be the same as on Rt1051
- Log headers were incomplete
- Log filtering was broken on linux
